Refactor lens model classes to use updated potential classes
